refactor(events): split label rendering from resource parsing
Two responsibilities, two classes:

- Appwrite\\Event\\Label\\Renderer::render() substitutes {namespace.key}
  placeholders inside a route label and returns the rendered string. Used
  by audit logging (api.php) and the cloud usage resolver to turn route
  labels into concrete resource paths.
- Appwrite\\Event\\Resource\\Parser parses that rendered path string into
  ordered {type,id} segments and exposes consumer-specific projections —
  auditTarget() for the granular audit-log shape (deepest segment + joined
  parent), and capForUsage() for the container-level usage cap.

Split the original combined Parser test into RendererTest and ParserTest.
